refactor(invitation): move the preview dialog to support
